add an environ variable to start docker or not

+++ linphonelib/linphonesession.py
@@ -123,7 +123,11 @@ audio_rtp_port={rtp_port}
         return cmd.execute(self._process)
 
     def _start(self):
-        cmd = "docker run --rm -ti -v {linphonerc}:/root/.linphonerc {docker_image}"
+        if os.getenv('USE_DOCKER'):
+            cmd = "docker run --rm -ti -v {linphonerc}:/root/.linphonerc {docker_image}"
+        else:
+            cmd = 'sh -c "linphonec -c {linphonerc}" &'
+
         config_file = self._create_config_file()
 
         self._process = pexpect.spawn(cmd.format(linphonerc=config_file,
